requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
astunparse==1.6.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.1
greenlet==3.0.3
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
mdtraj @ git+https://github.com/mdtraj/mdtraj.git@f134696b3d6e3ae26b00221394779edcfc00d898
numpy==1.26.4
packaging==24.0
parse==1.20.1
parse-type==0.6.2
playwright==1.44.0
pluggy==1.5.0
pyee==11.1.0
pyparsing==3.1.2
pytest==8.2.1
pytest-base-url==2.1.0
pytest-bdd==7.1.2
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-playwright==0.5.0
pytest-timeout==2.3.1
python-slugify==8.0.4
requests==2.32.3
scipy==1.13.1
six==1.16.0
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.12.1
urllib3==2.2.1
